finally made the userid pass through frontend to backend and made the payload structure correct
